CV/2021-03-10 JUnit/InheritanceExamples/src/cz/ou/kip/oopr2/filtering/filtering/MinimalSizeFilter.java
package cz.ou.kip.oopr2.filtering.filtering;

import cz.ou.kip.oopr2.filtering.Contracts;
import cz.ou.kip.oopr2.filtering.FileInfo;

public class MinimalSizeFilter extends Filter {
  public final long minimalSize;

  public MinimalSizeFilter(long minimalSize) {
    Contracts.isTrue(minimalSize >= 0, "Minimal size must be positive integer.");

    this.minimalSize = minimalSize;
  }

  @Override
  public boolean accepts(FileInfo fileInfo) {
    return fileInfo.getSizeInB() >= this.minimalSize;
  }
}
